Enhances UI and adds careers page
Improves the layout and spacing of several components (Connect, FAQ, Feature, Hero, Opinions, Testimonial, Unis) by increasing the maximum width for larger screens.

Adds a new Careers page that embeds a job listing from an external service.

Updates the Footer component to link to the external careers page and open the Devlog link in a new tab.

Enhances Navbar component with motion for the waitlist button and refactors to improve responsiveness.

Updates pricing page to improve layout and descriptions.